+++ src/main/java/com/consertreservation/domain/seat/model/SeatStatus.java
@@ -1,4 +1,4 @@
-package com.consertreservation.domain.seat;
+package com.consertreservation.domain.seat.model;
 
 public enum SeatStatus {
     RESERVED, AVAILABLE
